Move most of the code, refactor implementations
