+++ core/globals.cpp
@@ -35,6 +35,8 @@
 #include "bind/core_bind.h"
 #include "os/os.h"
 #include "io/file_access_pack.h"
+
+
 #include "io/file_access_network.h"
 
 Globals *Globals::singleton=NULL;
