+++ core/io/SCsub
@@ -4,7 +4,7 @@ env.add_source_files(env.core_sources,"*.cpp")
 env.add_source_files(env.core_sources,"*.c")
 #env.core_sources.append("io/fastlz.c")
 
-#TODO: Define SINGLE_PACK_SOURCE as an compile option, and then when it is selected, check that
+#TODO: Define USE_SINGLE_PACK_SOURCE as an compile option, and then when it is selected, check that
 #a *single* pack source is compiled
 
 env.add_source_files(env.core_sources,"formats/file_access_pck.cpp")
