+++ core/io/formats/file_access_zip.h
@@ -37,22 +37,15 @@
 #include "core/map.h"
 
 class ZipArchive : public PackSource {
-
 public:
-
 	struct File {
-
 		int package;
 		unz_file_pos file_pos;
 		File() {
-
 			package = -1;
 		};
 	};
-
-
 private:
-
 	struct Package {
 		String filename;
 		unzFile zfile;
@@ -64,9 +57,7 @@ private:
 	static ZipArchive* instance;
 
 	FileAccess::CreateFunc fa_create_func;
-
 public:
-
 	void close_handle(unzFile p_file) const;
 	unzFile get_file_handle(String p_file) const;
 
@@ -74,8 +65,8 @@ public:
 
 	bool file_exists(String p_name) const;
 
-	virtual bool try_open_pack(const String& p_path);
-	FileAccess* get_file(const String& p_path, PackedData::PackedFile* p_file);
+	virtual bool try_open_pack(const String& p_path, bool p_replace_files);
+	FileAccess* get_file(const String& p_path);
 
 	static ZipArchive* get_singleton();
 
@@ -85,16 +76,13 @@ public:
 
 
 class FileAccessZip : public FileAccess {
-
 	unzFile zfile;
 	unz_file_info64	file_info;
 
 	mutable bool at_eof;
 
 	ZipArchive* archive;
-
 public:
-
 	virtual Error _open(const String& p_path, int p_mode_flags); ///< open a file
 	virtual void close(); ///< close a file
 	virtual bool is_open() const; ///< true when file is open
@@ -116,7 +104,7 @@ public:
 
 	virtual uint64_t _get_modified_time(const String& p_file) { return 0; } // todo
 
-	FileAccessZip(const String& p_path, const PackedData::PackedFile& p_file);
+	FileAccessZip(const String& p_path);
 	~FileAccessZip();
 };
