+++ main/main.cpp
@@ -72,7 +72,9 @@
 #include "core/io/stream_peer_tcp.h"
 #include "core/os/thread.h"
 #include "core/io/file_access_pack.h"
-#include "core/io/file_access_zip.h"
+#include "core/io/formats/file_access_zip.h"
+#include "core/io/formats/file_access_pck.h"
+#include "core/io/formats/file_access_hpck.h"
 #include "translation.h"
 #include "version.h"
 
@@ -245,6 +247,9 @@ Error Main::setup(const char *execpath,int argc, char *argv[],bool p_second_phas
 	packed_data = PackedData::get_singleton();
 	if (!packed_data)
 		packed_data = memnew(PackedData);
+
+	//TODO: Put this line under ifdefs like the zip archive under it
+	packed_data->add_pack_source(PackedSourcePCK::get_singleton());
 
 #ifdef MINIZIP_ENABLED
 	packed_data->add_pack_source(ZipArchive::get_singleton());
